Removed rel="canonical", closes #3758
On the advice of the following articles:
  - http://googlewebmastercentral.blogspot.com/2013/04/5-common-mistakes-with-relcanonical.html
  - https://moz.com/blog/rel-confused-answers-to-your-rel-canonical-questions

rel="canonical" should not be shown on the same page
as rel="prev" and rel="next" as Google will implicitly
assume that they all point to the same page. With the
"pageless" variety only showing the first page of posts,
it explains exactly why any post content after the first
page is not indexed by Google.

... or perhaps it *is* indexed, but not returned. Who
the heck knows. :smile: