Updated 404 for image handling
If an image or asset specified in static file handler is not
found (ENOENT), then the 404 handler is now invoked (as opposed to
prior, where a handled exception was thrown).

Also, when requesting images inline that do not exist, NodeBB
will now send back "404 Not Found" instead of the entire 404
page. If you access the broken link directly, you'll see the
404 page.